Fix SQL execution in the app
The MSD app does not handle SQL execution from file very well at the moment.
The file is simply split by semicolons and, if the result is not an empty
string, it is fed into SQLite as a command. However, if single line comment
contains a semicolon, that comment my be passed to SQLite without any real
command. Apparently, SQLite does not like that.